\name{Gen_to_fas}
\alias{Gen_to_fas}
\title{Gen_to_fas}
\usage{
Gen_to_fas(genbank,seq_type)
}
\description{
Prints ID and sequence information in genbank file and output them
}
\arguments{
genbank: genbank(.gb) file
seq_type: sequence type(1:sigle sequence;2:multiple sequence)
}
\value{
return gene ID and sequence information(one item)
}
